Fixed prime specs so they work in ruby 1.8
The previous prime specs would only work in ruby 1.9+. In ruby 1.8 there was a prime class, but it was loaded differently ("require 'mathn'" rather than "require 'prime'") and had far fewer methods. I've updated the specs to load prime using "require 'mathn'" instead of "require 'prime'" (this works in 1.9 as well as 1.8), and to use a ruby version guard where appropriate. Now it works (as before) in 1.9, and doesn't error out in 1.8